Prepping for package refactor and header update
