+++ src/main/java/com/edwardawebb/stash/ssh/ao/EnterpriseKeyRepository.java
@@ -19,3 +19,4 @@ public interface EnterpriseKeyRepository {
     void removeRecord(SshKeyEntity key);
 
 }
+
