+++ src/test/java/it/com/edwardawebb/stash/ssh/EnterpriseSshKeyManagerImplFuncTest.java
@@ -3,7 +3,8 @@ package it.com.edwardawebb.stash.ssh;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import com.edwardawebb.stash.ssh.EnterpriseSshKeyServiceImpl;
+
+import com.edwardawebb.stash.ssh.keys.EnterpriseSshKeyServiceImpl;
 
 import static org.mockito.Mockito.*;
